排程候選商品不再自動建開團，改由商品編輯頁手動建立
- rpc_schedule_candidate：排程候選商品只建 draft product + sku（+ retail price）
  並標記候選 scheduled，移除自動建立 group_buy_campaigns / campaign_items 段落
  （新 migration 以 20260514000010 為基底擴寫，rollback 指回該版本）
- 抽出 CreateCampaignModal 為共用元件，商品列表頁沿用
- 商品編輯頁「關聯開團」面板新增「建立新開團」按鈕，
  走 rpc_create_campaign_from_product 手動開團
- 同步更新選品池頁註解
